Move relay tracing into shared
