+++ apps/mobile/src/features/observability/mobileTracing.test.ts
@@ -3,7 +3,8 @@ import * as Effect from "effect/Effect";
 import * as Layer from "effect/Layer";
 import { vi } from "vite-plus/test";
 
-import { remoteHttpClientLayer, withRelayClientTracing } from "@t3tools/client-runtime";
+import { remoteHttpClientLayer } from "@t3tools/client-runtime";
+import { withRelayClientTracing } from "@t3tools/shared/relayTracing";
 
 import { makeMobileTracingLayer } from "./mobileTracing";
