+++ apps/server/package.json
@@ -34,7 +34,6 @@
   },
   "devDependencies": {
     "@effect/vitest": "catalog:",
-    "@t3tools/client-runtime": "workspace:*",
     "@t3tools/contracts": "workspace:*",
     "@t3tools/shared": "workspace:*",
     "@t3tools/tailscale": "workspace:*",
