+++ apps/web/src/lib/runtime.ts
@@ -3,8 +3,9 @@ import type * as Effect from "effect/Effect";
 import * as Layer from "effect/Layer";
 import { FetchHttpClient } from "effect/unstable/http";
 
-import { makeRelayClientTracingLayer, remoteHttpClientLayer } from "@t3tools/client-runtime";
+import { remoteHttpClientLayer } from "@t3tools/client-runtime";
 import { httpHeaderRedactionLayer } from "@t3tools/shared/httpObservability";
+import { makeRelayClientTracingLayer } from "@t3tools/shared/relayTracing";
 import {
   PrimaryEnvironmentHttpClient,
   primaryEnvironmentHttpClientLive,
