+++ packages/client-runtime/src/index.ts
@@ -28,4 +28,3 @@ export * from "./checkpointDiffState.ts";
 export * from "./remote.ts";
 export * from "./managedRelay.ts";
 export * from "./managedRelayState.ts";
-export * from "./relayTracing.ts";
