+++ packages/shared/src/relayTracing.ts
@@ -20,7 +20,7 @@ export interface RelayClientTracingResource {
 }
 
 export class RelayClientTracer extends Context.Reference(
-  "@t3tools/client-runtime/relayTracing/RelayClientTracer",
+  "@t3tools/shared/relayTracing/RelayClientTracer",
   {
     defaultValue: () => Option.none<Tracer.Tracer>(),
   },
